consolidate typeahead files so that typeahead index is the input, there is no need to export the options, they are never used outside of the typeahead component. export as typeahead instead of typeaheadInput.
